coupon management after the pool has ended

+++ match-service/src/main/java/com/poorbet/matchservice/match/stream/controller/OddsController.java
@@ -7,6 +7,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Slf4j
@@ -18,7 +19,7 @@ public class OddsController {
     private final OddsService oddsService;
 
     @GetMapping("/{matchId}/odds")
-    public ResponseEntity<Double> getOdds(@PathVariable UUID matchId, @RequestParam OddsType type) {
+    public ResponseEntity<BigDecimal> getOdds(@PathVariable UUID matchId, @RequestParam OddsType type) {
         return oddsService.getOdds(matchId, type)
                 .map(ResponseEntity::ok)
                 .orElseGet(() -> ResponseEntity.notFound().build());
