+++ match-service/src/main/java/com/poorbet/matchservice/match/stream/projections/AwayWin.java
@@ -1,5 +1,7 @@
 package com.poorbet.matchservice.match.stream.projections;
 
+import java.math.BigDecimal;
+
 public interface AwayWin {
-    double getAwayWin();
+    BigDecimal getAwayWin();
 }
