+++ match-service/src/main/java/com/poorbet/matchservice/match/stream/projections/HomeWin.java
@@ -1,5 +1,7 @@
 package com.poorbet.matchservice.match.stream.projections;
 
+import java.math.BigDecimal;
+
 public interface HomeWin {
-    double getHomeWin();
+    BigDecimal getHomeWin();
 }
