+++ match-service/src/main/java/com/poorbet/matchservice/match/stream/service/OddsService.java
@@ -2,9 +2,10 @@ package com.poorbet.matchservice.match.stream.service;
 
 import com.poorbet.matchservice.match.stream.model.enums.OddsType;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 import java.util.UUID;
 
 public interface OddsService {
-    public Optional<Double> getOdds(UUID matchId, OddsType type);
+    public Optional<BigDecimal> getOdds(UUID matchId, OddsType type);
 }
